requirements/requirements.txt
# We use compatible release functionality (see PEP 440 here: https://www.python.org/dev/peps/pep-0440/#compatible-release)
# to specify acceptable version ranges of our project dependencies. This gives us the flexibility to keep up with small
# updates/fixes, whilst ensuring we don't install a major update which could introduce backwards incompatible changes.
# numpy>=1.21.0,<2.0.0

boto3>=1.35.34
uvicorn==0.18.3


fastapi==0.80.0
starlette==0.19.1
loguru>=0.5.3,<1.0.0
pillow>= 10.4.0
requests==2.28.1

python-multipart==0.0.10

numpy>=1.19.2,<2.0.0
pandas>=1.3.5,<2.0.0
pydantic>=1.8.1,<2.0.0
scikit-learn>=1.1.3,<2.0.0
strictyaml>=1.3.2,<2.0.0
ruamel.yaml>=0.17.4,<1.0.0
feature-engine>=1.0.2,<1.6.0
joblib>=1.0.1,<2.0.0
opencv-python>=4.5.1,<5.0.0
tqdm>=4.65.0,<5.0.0
# tensorflow>=2.5.0,<3.0.0
matplotlib==3.4.3

# tensorflow==2.10.0
tensorflow>=2.13.0,<=2.15.0
# tid-nn-model==0.0.1

mlflow>=2.17.0

mangum >= 0.19.0


# numpy>=1.19.2,<2.0.0
# pandas>=1.3.5,<2.0.0
# pydantic>=1.8.1,<2.0.0
# scikit-learn>=1.1.3,<2.0.0
# strictyaml>=1.3.2,<2.0.0
# ruamel.yaml>=0.17.4,<1.0.0
# feature-engine>=1.0.2,<1.6.0
# joblib>=1.0.1,<2.0.0
# opencv-python>=4.5.1,<5.0.0
# tqdm>=4.65.0,<5.0.0
# # tensorflow>=2.5.0,<3.0.0
# matplotlib==3.4.3

# # tensorflow==2.10.0
# tensorflow>=2.13.0,<=2.15.0
